Install built-in commands during initialization
