+++ src/wasm-terminal/command/callback-command.ts
