+++ src/wasm-terminal/command/command-options.ts
