+++ src/wasm-terminal/command/command.ts
