+++ src/wasm-terminal/command/go-wasm-command.ts
