+++ src/wasm-terminal/command/go/go.d.ts
