+++ src/wasm-terminal/command/go/go.js
