+++ src/wasm-terminal/command/go/polyfill.d.ts
