+++ src/wasm-terminal/command/go/polyfill.js
