+++ src/wasm-terminal/command/wasi-command.ts
