+++ src/wasm-terminal/functions/fetch-command-wapm.ts
