+++ src/wasm-terminal/wasm-shell/shell-history.ts
