+++ src/wasm-terminal/wasm-shell/shell-utils.ts
