+++ src/wasm-terminal/wasm-shell/wasm-shell.ts
